chore(deps): upgrade grpc-java from 1.83.0 to 1.83.1
1. bump grpcVersion to 1.83.1 to pick up the upstream fix for
   grpc/grpc-java#12930 (PR grpc/grpc-java#12942), which enforces
   connection.remote().maxActiveStreams(maxStreams) at handler startup
2. drop GrpcNettyMaxConcurrentStreamsLimiter, the local protocol-negotiator
   shim that applied the same limit while 1.83.0 left the remote endpoint
   unbounded until the client acknowledged SETTINGS
3. set maxConcurrentCallsPerConnection directly on NettyServerBuilder in
   RpcService, as it was done before the shim
4. move shouldIgnoreClientMaxHeaderListSizeOnServer from the deleted shim
   test into NettyHttp2HeaderSecurityTest, since it guards netty
   CVE-2026-50560 rather than the shim
5. update the io.grpc checksums in gradle/verification-metadata.xml
6. keep RpcServiceHttp2SecurityTest as the regression guard: it fails on
   1.83.0 without the shim and passes on 1.83.1
